Add tests for new permission-target all authenticated users
Add two tests to check that permissions on all authenticated users have less
precedence (and do work) than group/specific-user permissions.
In shortcuts.py we edit assign_perm and remove_perm to build a specific query
when the target authenticated_users is used.

+++ machina/apps/forum_permission/shortcuts.py
@@ -17,42 +17,76 @@ ForumPermission = get_model('forum_permission', 'ForumPermission')
 GroupForumPermission = get_model('forum_permission', 'GroupForumPermission')
 UserForumPermission = get_model('forum_permission', 'UserForumPermission')
 
+ALL_AUTH = object()  # object to check against if all authenticated users are meant
+
 
 class NotUserNorGroup(Exception):
     pass
 
 
-def assign_perm(perm, user_or_group, forum=None, has_perm=True):
-    """ Assigns a permission to a user (anonymous or not) or a group. """
-    user, group = get_identity(user_or_group)
+def assign_perm(perm, object, forum=None, has_perm=True):
+    """ Assigns a permission to a user (anonymous, specific or all authenticated) or a group.
+    object can be either: a user object (specific or anonymous)
+                          a group object
+                          the ALL_AUTH object for all authenticated users
+    """
     perm = ForumPermission.objects.get(codename=perm)
-    if user:
+
+    if object is ALL_AUTH:
         return UserForumPermission.objects.create(
             forum=forum,
             permission=perm,
-            user=user if not user.is_anonymous else None,
-            anonymous_user=user.is_anonymous,
+            user=None,
+            anonymous_user=False,
+            authenticated_user=True,
             has_perm=has_perm,
         )
-    if group:
-        return GroupForumPermission.objects.create(
-            forum=forum, permission=perm, group=group, has_perm=has_perm,
-        )
+    else:
+        user, group = get_identity(object)
+
+        if user:
+            return UserForumPermission.objects.create(
+                forum=forum,
+                permission=perm,
+                user=user if not user.is_anonymous else None,
+                anonymous_user=user.is_anonymous,
+                authenticated_user=False,
+                has_perm=has_perm,
+            )
+        if group:
+            return GroupForumPermission.objects.create(
+                forum=forum, permission=perm, group=group, has_perm=has_perm,
+            )
 
 
-def remove_perm(perm, user_or_group, forum=None):
-    """ Remove a permission to a user (anonymous or not) or a group. """
-    user, group = get_identity(user_or_group)
+def remove_perm(perm, object, forum=None):
+    """ Remove a permission to a user (anonymous, specific or all authenticated) or a group.
+    object can be either: a user object (specific or anonymous)
+                          a group object
+                          the ALL_AUTH object for all authenticated users
+    """
     perm = ForumPermission.objects.get(codename=perm)
-    if user:
-        UserForumPermission.objects.filter(
+
+    if object is ALL_AUTH:
+        return UserForumPermission.objects.filter(
             forum=forum,
             permission=perm,
-            user=user if not user.is_anonymous else None,
-            anonymous_user=user.is_anonymous,
+            user=None,
+            anonymous_user=False,
+            authenticated_user=True,
         ).delete()
-    if group:
-        GroupForumPermission.objects.filter(forum=forum, permission=perm, group=group).delete()
+    else:
+        user, group = get_identity(object)
+        if user:
+            UserForumPermission.objects.filter(
+                forum=forum,
+                permission=perm,
+                user=user if not user.is_anonymous else None,
+                anonymous_user=user.is_anonymous,
+                authenticated_user=False,
+            ).delete()
+        if group:
+            GroupForumPermission.objects.filter(forum=forum, permission=perm, group=group).delete()
 
 
 def get_identity(identity):
